Project structure refactor.
