+++ api/echo.go_
@@ -12,3 +12,54 @@ func EchoHandleFunc(w http.ResponseWriter, r *http.Request) {
 	w.Header().Add("Content-Type", "text/plain")
 	fmt.Fprintf(w, message)
 }*/
+
+// mathHandler does a math operation
+func mathHandler(w http.ResponseWriter, r *http.Request) {
+	// Decode request
+	defer r.Body.Close()
+	dec := json.NewDecoder(r.Body)
+	req := &MathRequest{}
+
+	if err := dec.Decode(req); err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	// Do work
+	resp := &MathResponse{}
+	switch req.Op {
+	case "+":
+		resp.Result = req.Left + req.Right
+	case "-":
+		resp.Result = req.Left - req.Right
+	case "*":
+		resp.Result = req.Left * req.Right
+	case "/":
+		if req.Right == 0.0 {
+			resp.Error = "division by 0"
+		} else {
+			resp.Result = req.Left / req.Right
+		}
+	default:
+		resp.Error = fmt.Sprintf("unknown operation: %s", req.Op)
+	}
+
+	// Encode & return result
+	w.Header().Set("Content-Type", "application/json")
+	if resp.Error != "" {
+		w.WriteHeader(http.StatusBadRequest)
+	}
+
+	enc := json.NewEncoder(w)
+	if err := enc.Encode(resp); err != nil {
+		// Can't return error to client here
+		log.Printf("can't encode %v - %s", resp, err)
+	}
+}
+func main() {
+	http.HandleFunc("/hello", helloHandler)
+	http.HandleFunc("/math", mathHandler)
+	if err := http.ListenAndServe(":8080", nil); err != nil {
+		log.Fatal(err)
+	}
+}
