+++ api/hello.go_
@@ -23,3 +23,8 @@ func HelloHandleFunc(w http.ResponseWriter, r *http.Request) {
 	w.Header().Add("Content-Type", "application/json; charset=utf-8")
 	w.Write(b)
 }*/
+
+	resp := map[string]interface{}{
+		"ok":      true,
+		"balance": prevBalance + req.Amount,
+	}
