--- sensors/github_api.go_
@@ -1,44 +1,0 @@
-/*package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"log"
-	"net/http"
-)
-
-// User is a github user information
-type User struct {
-	Name        string `json:"name"`
-	PublicRepos int    `json:"public_repos"`
-}
-
-// userInfo return information on github user
-func userInfo(login string) (*User, error) {
-	// HTTP call
-	url := fmt.Sprintf("https://api.github.com/users/%s", login)
-	resp, err := http.Get(url)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	// Decode JSON
-	user := &User{}
-	dec := json.NewDecoder(resp.Body)
-	if err := dec.Decode(user); err != nil {
-		return nil, err
-	}
-
-	return user, nil
-}
-
-func main() {
-	user, err := userInfo("tebeka")
-	if err != nil {
-		log.Fatalf("error: %s", err)
-	}
-
-	fmt.Printf("%+v\n", user)
-}
-*/
