--- sensors/json_complete.go_
@@ -1,52 +1,0 @@
-// JSON example
-/*package main
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"log"
-	"os"
-)
-
-var data = `
-{
-  "user": "Scrooge McDuck",
-  "type": "deposit",
-  "amount": 1000000.3
-}
-`
-
-// Request is a bank transactions
-type Request struct {
-	Login  string  `json:"user"`
-	Type   string  `json:"type"`
-	Amount float64 `json:"amount"`
-}
-
-func main() {
-	rdr := bytes.NewBufferString(data) // Simulate a file/socket
-
-	// Decode request
-	dec := json.NewDecoder(rdr)
-
-	req := &Request{}
-	if err := dec.Decode(req); err != nil {
-		log.Fatalf("error: can't decode - %s", err)
-	}
-
-	fmt.Printf("got: %+v\n", req)
-
-	// Create response
-	prevBalance := 8500000.0 // Loaded from database
-	resp := map[string]interface{}{
-		"ok":      true,
-		"balance": prevBalance + req.Amount,
-	}
-
-	// Encode respose
-	enc := json.NewEncoder(os.Stdout)
-	if err := enc.Encode(resp); err != nil {
-		log.Fatalf("error: can't encode - %s", err)
-	}
-}*/
