--- sensors/req.json
@@ -1,4 +1,0 @@
-{
-  "key": "x",
-  "value": 1
-}
